requirements.txt
langchain
openai
huggingface_hub
python_dotenv
streamlit
response